Refactor tests to test each delegate
